move landsat and planet to components
