+++ components/basemaps/landsat-menu/styles.scss
@@ -1,0 +1,4 @@
+@import '~styles/settings.scss';
+
+.c-landsat-menu {
+}
