+++ components/basemaps/planet-menu/styles.scss
@@ -1,0 +1,4 @@
+@import '~styles/settings.scss';
+
+.c-planet-menu {
+}
